Add publisher to game info
