+++ app/scraper/info/utils.py
@@ -1,5 +1,6 @@
 import difflib
 import re
+from typing import Any
 
 RESULT_MATCH_THRESHOLD = 0.85
 
@@ -43,3 +44,16 @@ def get_match_score(search: str, result: str) -> float:
         score -= 0.01
 
     return score
+
+
+def clean_nones(value: dict[str, Any]) -> dict[str, Any]:
+    """
+    Recursively remove all None values from dictionaries and lists, and returns
+    the result as a new dictionary or list.
+    """
+    if isinstance(value, list):
+        return [clean_nones(x) for x in value if x is not None]
+    elif isinstance(value, dict):
+        return {key: clean_nones(val) for key, val in value.items() if val is not None}
+    else:
+        return value
